Moved CouponAdapter.java
